+++ app/src/main/java/me/alberto/cowrywiseconveter/data/remote/model/Query.kt
@@ -3,5 +3,5 @@ package me.alberto.cowrywiseconveter.data.remote.model
 data class Query (
     val from: String,
     val to: String,
-    val amount: Long
+    val amount: Double
 )
