+++ app/src/main/java/me/alberto/cowrywiseconveter/data/remote/source/IRemoteSource.kt
@@ -2,5 +2,5 @@ package me.alberto.cowrywiseconveter.data.remote.source
 
 interface IRemoteSource {
     suspend fun getSymbols(): Result<List<String>>
-    suspend fun convert(from: String, to: String, amount: Long): Result<Double>
+    suspend fun convert(from: String, to: String, amount: Double): Result<Double>
 }
